Refactor Duplicate Loops to Single
Domain: It contains the Log In Monitor screen which displays Stats like total Logins, Unique Users and also mentions the same info for today, this month, last month and 2 months ago. It also displays all the users, number of times they have logged In and the latest date they have logged In. 

Problem Statement: There is a duplicate code in the loops that runs for this month, last month and 2 months ago. Also there is a bug in the same stats.